Fix spelling errors (#11)
This might be a bit controversial, since we like to keep things as is in
our community. However I feel when this is document is what we legally
follow, that it is a good idea to fix these things.
I am uncertain what 'papirsvirke' should be, since the only thing that
google shows when googling the word is our 'vedtægter'. I think I
corrected it properly.